backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.1.1
sqlparse==0.5.3
